agent_economy/discord_monitor/requirements.txt
discord.py-self>=2.1.0
aiohttp>=3.9.0
